test_requirements.txt
# Test dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
factory-boy==3.3.0
faker==20.1.0
